requirements.txt
pandas
numpy
scikit-learn
joblib
google-cloud-aiplatform
